refactor: ``backup``, ``list_backups`` and ``retore_table`` methods moved to the ``Table`` class.
