+++ tests/unit/test_table.py
@@ -128,8 +128,12 @@ class TestTable(unittest.TestCase):
     PROJECT_ID = "project-id"
     INSTANCE_ID = "instance-id"
     INSTANCE_NAME = "projects/" + PROJECT_ID + "/instances/" + INSTANCE_ID
+    CLUSTER_ID = "cluster-id"
+    CLUSTER_NAME = INSTANCE_NAME + "/clusters/" + CLUSTER_ID
     TABLE_ID = "table-id"
     TABLE_NAME = INSTANCE_NAME + "/tables/" + TABLE_ID
+    BACKUP_ID = "backup-id"
+    BACKUP_NAME = CLUSTER_NAME + "/backups/" + BACKUP_ID
     ROW_KEY = b"row-key"
     ROW_KEY_1 = b"row-key-1"
     ROW_KEY_2 = b"row-key-2"
@@ -1152,6 +1156,158 @@ class TestTable(unittest.TestCase):
         table_api.test_iam_permissions.assert_called_once_with(
             resource=table.name, permissions=permissions
         )
+
+    def test_backup_factory_defaults(self):
+        from google.cloud.bigtable.backup import Backup
+
+        instance = self._make_one(self.INSTANCE_ID, None)
+        table = self._make_one(self.TABLE_ID, instance)
+        backup = table.backup(self.BACKUP_ID)
+
+        self.assertIsInstance(backup, Backup)
+        self.assertEqual(backup.backup_id, self.BACKUP_ID)
+        self.assertIs(backup._instance, instance)
+        self.assertIsNone(backup._cluster)
+        self.assertEqual(backup.table_id, self.TABLE_ID)
+        self.assertIsNone(backup._expire_time)
+
+        self.assertIsNone(backup._parent)
+        self.assertIsNone(backup._source_table)
+        self.assertIsNone(backup._start_time)
+        self.assertIsNone(backup._end_time)
+        self.assertIsNone(backup._size_bytes)
+        self.assertIsNone(backup._state)
+
+    def test_backup_factory_non_defaults(self):
+        import datetime
+        from google.cloud._helpers import UTC
+        from google.cloud.bigtable.backup import Backup
+
+        instance = self._make_one(self.INSTANCE_ID, None)
+        table = self._make_one(self.TABLE_ID, instance)
+        timestamp = datetime.datetime.utcnow().replace(tzinfo=UTC)
+        backup = table.backup(
+            self.BACKUP_ID,
+            cluster_id=self.CLUSTER_ID,
+            expire_time=timestamp,
+        )
+
+        self.assertIsInstance(backup, Backup)
+        self.assertEqual(backup.backup_id, self.BACKUP_ID)
+        self.assertIs(backup._instance, instance)
+
+        self.assertEqual(backup.backup_id, self.BACKUP_ID)
+        self.assertIs(backup._cluster, self.CLUSTER_ID)
+        self.assertEqual(backup.table_id, self.TABLE_ID)
+        self.assertEqual(backup._expire_time, timestamp)
+        self.assertIsNone(backup._start_time)
+        self.assertIsNone(backup._end_time)
+        self.assertIsNone(backup._size_bytes)
+        self.assertIsNone(backup._state)
+
+    def _list_backups_helper(self, cluster_id=None, filter_=None, **kwargs):
+        from google.cloud.bigtable_admin_v2.gapic import (
+            bigtable_instance_admin_client,
+            bigtable_table_admin_client,
+        )
+        from google.cloud.bigtable_admin_v2.proto import (
+            bigtable_table_admin_pb2,
+            table_pb2,
+        )
+        from google.cloud.bigtable.backup import Backup
+
+        instance_api = bigtable_instance_admin_client.BigtableInstanceAdminClient
+        table_api = bigtable_table_admin_client.BigtableTableAdminClient(mock.Mock())
+        client = self._make_client(
+            project=self.PROJECT_ID, credentials=_make_credentials(), admin=True
+        )
+        instance = client.instance(instance_id=self.INSTANCE_ID)
+        table = self._make_one(self.TABLE_ID, instance)
+
+        client._instance_admin_client = instance_api
+        client._table_admin_client = table_api
+
+        parent = self.INSTANCE_NAME + "/clusters/cluster/"
+        backups_pb = bigtable_table_admin_pb2.ListBackupsResponse(
+            backups=[
+                table_pb2.Backup(name=parent + "/backups/op1"),
+                table_pb2.Backup(name=parent + "/backups/op2"),
+                table_pb2.Backup(name=parent + "/backups/op3"),
+            ]
+        )
+
+        api = table_api._inner_api_calls["list_backups"] = mock.Mock(
+            return_value=backups_pb
+        )
+
+        backups = table.list_backups(
+            cluster_id=cluster_id, filter_=filter_, **kwargs
+        )
+
+        for backup in backups:
+            self.assertIsInstance(backup, Backup)
+
+        if not cluster_id:
+            cluster_id = "-"
+        parent = "{}/clusters/{}".format(self.INSTANCE_NAME, cluster_id)
+
+        expected_metadata = [
+            ("x-goog-request-params", "parent={}".format(parent)),
+        ]
+        api.assert_called_once_with(
+            bigtable_table_admin_pb2.ListBackupsRequest(
+                parent=parent, filter=filter_, **kwargs
+            ),
+            retry=mock.ANY,
+            timeout=mock.ANY,
+            metadata=expected_metadata,
+        )
+
+    def test_list_backups_defaults(self):
+        self._list_backups_helper()
+
+    def test_list_backups_w_options(self):
+        self._list_backups_helper(
+            cluster_id="cluster",
+            filter_="filter",
+            order_by="order_by",
+            page_size=10
+        )
+
+    def _restore_helper(self, backup_name=None):
+        from google.cloud.bigtable_admin_v2 import BigtableTableAdminClient
+        from google.cloud.bigtable_admin_v2.gapic import bigtable_instance_admin_client
+        from google.cloud.bigtable.instance import Instance
+
+        op_future = object()
+        instance_api = bigtable_instance_admin_client.BigtableInstanceAdminClient
+
+        client = mock.Mock(project=self.PROJECT_ID, instance_admin_client=instance_api)
+        instance = Instance(self.INSTANCE_ID, client=client)
+        table = self._make_one(self.TABLE_ID, instance)
+
+        api = client.table_admin_client = mock.create_autospec(
+            BigtableTableAdminClient, instance=True
+        )
+        api.restore_table.return_value = op_future
+
+        if backup_name:
+            future = table.restore(self.TABLE_ID, backup_name=self.BACKUP_NAME)
+        else:
+            future = table.restore(
+                self.TABLE_ID, self.CLUSTER_ID, self.BACKUP_ID
+            )
+        self.assertIs(future, op_future)
+
+        api.restore_table.assert_called_once_with(
+            parent=self.INSTANCE_NAME, table_id=self.TABLE_ID, backup=self.BACKUP_NAME,
+        )
+
+    def test_restore_table_w_backup_id(self):
+        self._restore_helper()
+
+    def test_restore_table_w_backup_name(self):
+        self._restore_helper(backup_name=self.BACKUP_NAME)
 
 
 class Test__RetryableMutateRowsWorker(unittest.TestCase):
